Add a basic test to find undefined names in code
There are a lot of potential bugs in Zenoss code, which can be found with
the static code analyser, so there is a sense to add this check to tests
and take care, that this test pass.

Added a simple proof-of-concept test, which shows undefined variables
in code. Used pylint [0] as a static code analyser and tox [1] as a
test runner.

[0] - http://www.pylint.org/
[1] - https://tox.readthedocs.org